move id_generator to aws crate
