+++ opentelemetry-aws/Cargo.toml
@@ -20,11 +20,12 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
 default = ["trace"]
-trace = ["opentelemetry/trace"]
+trace = ["opentelemetry/trace", "opentelemetry_sdk"]
 
 [dependencies]
 once_cell = "1.12"
 opentelemetry = { version = "0.21" }
+opentelemetry_sdk = { version = "0.21", optional = true }
 
 [dev-dependencies]
 opentelemetry_sdk = { version = "0.21", features = ["trace", "testing"] }
