+++ opentelemetry-aws/src/trace/mod.rs
@@ -1,0 +1,11 @@
+
+pub mod id_generator;
+pub mod xray_propagator;
+
+#[cfg(feature = "trace")]
+pub use xray_propagator::XrayPropagator;
+
+
+#[cfg(feature = "trace")]
+pub use id_generator::XrayIdGenerator;
+
